BPM-71 skills are working now
