+++ package.json
@@ -14,16 +14,17 @@
   "dependencies": {
     "@material-ui/core": "^3.3.1",
     "@material-ui/icons": "^3.0.1",
+    "axios": "latest",
     "mui-datatables": "^2.0.0-beta-53",
     "prop-types": "15.6.2",
     "react": "^16.6.0",
     "react-dom": "^16.6.0",
     "react-redux": "^5.0.7",
+    "react-router-dom": "^4.3.1",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
     "typeface-roboto": "0.0.54",
-    "webpack-dev-middleware": "^3.4.0",
-    "axios": "latest"
+    "webpack-dev-middleware": "^3.4.0"
   },
   "repository": {
     "type": "git",
