+++ src/AppConstants.js
@@ -1,0 +1,3 @@
+export const AppConstants = {
+  APP_TITLE: 'Here are all skills currently saved in our system',
+};
