+++ src/RootReducer.js
@@ -2,9 +2,9 @@ import { combineReducers } from 'redux';
 import skillDelete from './component/delete-dialog/DeleteDialogReducers';
 import skillList from './component/skills/SkillReducer';
 import hover from './component/hover/HoverReducer';
-import message from './component/notification/NotificationReducer';
+import message from './component/bpm-notification/NotificationReducer';
 import skillEdit from './component/skills/edit/EditSkillReducer';
-import inputError from './component/text-field/BpmTextFieldReducer';
+import inputError from './component/bpm-text-field/BpmTextFieldReducer';
 import categoryList from './component/categories/CategoriesReducer';
 
 const rootReducer = combineReducers({
