+++ src/component/bpm-buttons/BpmIconButtonConstants.js
@@ -1,0 +1,4 @@
+export const TooltipConst = {
+  ENTER_DELAY: 400,
+  LEAVE_DELAY: 200,
+};
