+++ src/component/bpm-footer/FooterConstants.js
@@ -1,0 +1,4 @@
+export const FooterConst = {
+  LEARN_MORE: 'Learn more about bpm',
+  BPM_GITHUB: 'https://github.com/ioet/',
+};
