+++ src/component/bpm-footer/FooterStyles.js
@@ -1,0 +1,10 @@
+export const FooterStyles = theme => ({
+  root: {
+    width: '90%',
+    margin: 'auto',
+    padding: theme.spacing.unit * 1,
+  },
+  button: {
+    margin: theme.spacing.unit * 1,
+  },
+});
