+++ src/component/bpm-notification/NotificationConstants.js
@@ -1,0 +1,23 @@
+export const NotificationConst = {
+  AUTO_HIDE_DURATION: 3000,
+};
+
+export const NotificationMessage = {
+  SKILL_CREATED_SUCCESSFULLY: ' was saved successfully.',
+  CHANGES_DISCARDED: 'Your changes have been discarded.',
+  CHANGES_UPDATED_SUCCESSFULLY: 'Your changes have been submitted successfully.',
+  SKILL_DELETED_SUCCESSFULLY: ' has been deleted successfully.',
+};
+
+export const PromptMessage = {
+  ENTER_VALID_NAME: 'Please enter a valid name.',
+};
+
+export const ErrorMessage = {
+  FAILED_TO_LOAD_SKILLS: 'Failed to load all saved skills.',
+  FAILED_TO_CREATE_SKILL: 'Failed to create a new skill.',
+  FAILED_TO_UPDATE_SKILL: 'Failed to update the skill.',
+  FAILED_TO_REMOVE_SKILL: 'Failed to deleteSkill the skill.',
+
+  FAILED_TO_LOAD_CATEGORIES: 'Failed to load all saved categories.',
+};
