+++ src/component/bpm-notification/NotificationStyles.js
@@ -1,0 +1,7 @@
+export const NotificationStyles = theme => ({
+  snackbar: {
+    [theme.breakpoints.up('md')]: {
+      marginBottom: theme.spacing.unit * 3,
+    },
+  },
+});
