+++ src/component/bpm-table-cell/BpmTableCell.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import withStyles from '@material-ui/core/styles/withStyles';
-import { TableCellStyles } from '../../styles';
+import { TableCellStyles } from './BpmTableCellStyles';
 
-const MyTableCell = (props) => {
+const BpmTableCell = (props) => {
   const {
     classes, onMouseOver, onMouseOut, children,
   } = props;
@@ -27,11 +27,15 @@ const MyTableCell = (props) => {
   );
 };
 
-MyTableCell.propTypes = {
+BpmTableCell.propTypes = {
   classes: PropTypes.object.isRequired,
   onMouseOver: PropTypes.func.isRequired,
   onMouseOut: PropTypes.func.isRequired,
-  children: PropTypes.element.isRequired,
+  children: PropTypes.oneOfType([
+    PropTypes.element,
+    PropTypes.string,
+    PropTypes.number,
+  ]).isRequired,
 };
 
-export default withStyles(TableCellStyles)(MyTableCell);
+export default withStyles(TableCellStyles)(BpmTableCell);
