+++ src/component/bpm-table-cell/BpmTableCellStyles.js
@@ -1,0 +1,10 @@
+export const TableCellStyles = theme => ({
+  cell: {
+    height: 'auto',
+    lineHeight: theme.spacing.unit / 2,
+    verticalAlign: 'middle',
+    paddingTop: theme.spacing.unit / 2,
+    paddingBottom: theme.spacing.unit / 2,
+    paddingLeft: theme.spacing.unit * 3,
+  },
+});
