+++ src/component/bpm-text-field/BpmTextFieldReducer.js
