+++ src/component/bpm-text-field/BpmTextFieldSelector.js
