--- src/component/notification/NotificationConsts.js
@@ -1,5 +1,0 @@
-const NotificationConst = {
-  AUTO_HIDE_DURATION: 3000,
-};
-
-export default NotificationConst;
