+++ src/component/skills/edit/EditSkillActions.js
@@ -1,11 +1,12 @@
 import { getCurrentEditSkill, isSkillCreation } from './EditSkillSelector';
 import SkillsApi from '../../skillsApi/SkillsApi';
-import { showMessage } from '../../notification/NotificationActions';
+import { showMessage } from '../../bpm-notification/NotificationActions';
 import { FAILED_TO_CREATE_SKILL } from '../create/NewSkillConstants';
 import { addSkills, updateSkillInList } from '../SkillActions';
 import { getSkillById } from '../SkillSelector';
-import { removeAllInputErrors } from '../../text-field/BpmTextFieldActions';
-import { ErrorMessage, NotificationMessage } from '../../../constants';
+import { removeAllInputErrors, setInputError } from '../../bpm-text-field/BpmTextFieldActions';
+import { ErrorMessage, NotificationMessage, PromptMessage } from '../../bpm-notification/NotificationConstants';
+import { SkillFormDialogNames } from '../dialog-form/SkillFormDialogConstants';
 
 export const EditSkillAction = {
   EDIT_START: 'EDIT_START',
@@ -37,9 +38,24 @@ export const stopEditSkill = () => ({
   type: EditSkillAction.EDIT_END,
 });
 
+const isValidField = input => typeof input !== 'undefined' && input !== '';
+
+const isInputValidSkill = (dispatch, skill) => {
+  if (!isValidField(skill.name)) {
+    dispatch(showMessage(PromptMessage.ENTER_VALID_NAME));
+    dispatch(setInputError(SkillFormDialogNames.SKILL_NAME));
+    return false;
+  }
+  return true;
+};
+
 export const createSkill = skill => (
   (dispatch) => {
-    new SkillsApi().createSkill(skill)
+    if (!isInputValidSkill(dispatch, skill)) {
+      return null;
+    }
+
+    return new SkillsApi().createSkill(skill)
       .then((response) => {
         dispatch(stopEditSkill());
         dispatch(addSkills([response.data]));
@@ -52,7 +68,11 @@ export const createSkill = skill => (
 );
 
 export const updateSkill = skill => ((dispatch) => {
-  new SkillsApi().updateSkill(skill)
+  if (!isInputValidSkill(dispatch, skill)) {
+    return null;
+  }
+
+  return new SkillsApi().updateSkill(skill)
     .then((response) => {
       dispatch(removeAllInputErrors());
       dispatch(stopEditSkill());
