+++ src/component/skills/list/BpmDeleteButtonContainer.jsx
@@ -1,15 +1,16 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Delete } from '@material-ui/icons';
-import BpmIconButton from './BpmIconButton';
-import { getHoverId, isHoverActive } from '../hover/HoverSelector';
-import { showDeleteDialog } from '../delete-dialog/DeleteDialogActions';
+import BpmIconButton from '../../bpm-buttons/BpmIconButton';
+import { getHoverId, isHoverActive } from '../../hover/HoverSelector';
+import { showDeleteDialog } from '../../delete-dialog/DeleteDialogActions';
+import { DELETE } from './SkillListConstants';
 
 const mapStateToProps = state => ({
   hover: isHoverActive(state),
   hoverId: getHoverId(state),
   icon: <Delete />,
-  tooltip: 'Delete',
+  tooltip: DELETE,
 });
 
 const mapDispatchToProps = (dispatch, ownProps) => ({
