+++ src/component/skills/list/BpmEditButtonContainer.jsx
@@ -1,15 +1,16 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Edit } from '@material-ui/icons';
-import { getHoverId, isHoverActive } from '../hover/HoverSelector';
-import BpmIconButton from './BpmIconButton';
-import { editSkill } from '../skills/edit/EditSkillActions';
+import { getHoverId, isHoverActive } from '../../hover/HoverSelector';
+import BpmIconButton from '../../bpm-buttons/BpmIconButton';
+import { editSkill } from '../edit/EditSkillActions';
+import { EDIT } from './SkillListConstants';
 
 const mapStateToProps = state => ({
   hover: isHoverActive(state),
   hoverId: getHoverId(state),
   icon: <Edit />,
-  tooltip: 'Edit',
+  tooltip: EDIT,
 });
 
 const mapDispatchToProps = (dispatch, ownProps) => ({
