+++ src/component/skills/list/SkillListConstants.js
@@ -1,0 +1,14 @@
+export const EDIT = 'Edit';
+
+export const DELETE = 'Delete';
+
+export const SkillListConst = {
+  COLUMN_ID: 'Id',
+  COLUMN_NAME: 'Name',
+  COLUMN_LABEL: 'Label',
+  COLUMN_CATEGORY: 'Category',
+  COLUMN_BUSINESS_VALUE: 'Business Value',
+  COLUMN_PREDICTIVE_VALUE: 'Predictive Value',
+  COLUMN_EDIT: 'Edit',
+  COLUMN_DELETE: 'Delete',
+};
