--- src/constants.js
@@ -1,61 +1,0 @@
-export const Variable = {
-  NAME: 'name',
-  PREDICTIVE_VALUE: 'predictive_value',
-  BUSINESS_VALUE: 'business_value',
-  CATEGORY: 'category_id',
-};
-
-export const ErrorMessage = {
-  FAILED_TO_LOAD_SKILLS: 'Failed to load all saved skills.',
-  FAILED_TO_CREATE_SKILL: 'Failed to create a new skill.',
-  FAILED_TO_UPDATE_SKILL: 'Failed to update the skill.',
-  FAILED_TO_REMOVE_SKILL: 'Failed to deleteSkill the skill.',
-
-  FAILED_TO_LOAD_CATEGORIES: 'Failed to load all saved categories.',
-};
-
-export const NotificationMessage = {
-  SKILL_CREATED_SUCCESSFULLY: ' was saved successfully.',
-  CHANGES_DISCARDED: 'Your changes have been discarded.',
-  CHANGES_UPDATED_SUCCESSFULLY: 'Your changes have been submitted successfully.',
-  SKILL_DELETED_SUCCESSFULLY: ' has been deleted successfully.',
-};
-
-export const PromptMessage = {
-  ENTER_VALID_NAME: 'Please enter a valid name.',
-};
-
-export const AppConst = {
-  APP_TITLE: 'Here are all skills currently saved in our system',
-};
-
-export const FooterConst = {
-  LEARN_MORE: 'Learn more about bpm',
-  BPM_GITHUB: 'https://github.com/ioet/',
-};
-
-export const SkillListItemConst = {
-  TOOLTIP_EDIT: 'Edit',
-  TOOLTIP_SAVE: 'Save',
-  TOOLTIP_DISCARD: 'Discard',
-  TOOLTIP_DELETE: 'Delete',
-  EDIT_NAME: 'Edit your name',
-  EDIT_BUSINESS_VALUE: 'Edit the business value',
-  EDIT_PREDICTIVE_VALUE: 'Edit the predictive value',
-};
-
-export const SkillListConst = {
-  COLUMN_ID: 'Id',
-  COLUMN_NAME: 'Name',
-  COLUMN_LABEL: 'Label',
-  COLUMN_CATEGORY: 'Category',
-  COLUMN_BUSINESS_VALUE: 'Business Value',
-  COLUMN_PREDICTIVE_VALUE: 'Predictive Value',
-  COLUMN_EDIT: 'Edit',
-  COLUMN_DELETE: 'Delete',
-};
-
-export const TooltipConst = {
-  ENTER_DELAY: 400,
-  LEAVE_DELAY: 200,
-};
