+++ src/index.js
@@ -9,9 +9,9 @@ import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider';
 import 'typeface-roboto';
 import App from './App';
 import rootReducer from './RootReducer';
-import { RootTheme } from './styles';
 import { getAllSkills } from './component/skills/SkillActions';
 import { getAllCategories } from './component/categories/CategoriesActions';
+import RootTheme from './AppStyles';
 
 const store = createStore(
   rootReducer,
